Fix aggregation logic while running tests locally
As part of b/331696130 the iteration renaming logic has been changed to rename the test method name, rather than the class name (so from `Class$1#method` to `Class#method$1`). Because of this, while grouping test cases for aggregation, additional checks need to happen to ensure that if the method name contains the iteration separator(as part of the above mentioned change), then it needs to be appropriately trimmed out so that grouping happens with the correct test id.

Bug: 342206870

Test: make catbox && make gcatbox
These changes have been manually tested by executing the test on a
locally connected reference device

Flag: TEST_ONLY

Change-Id: I8ef982eccf884e5db2c371d62d86097e67a4d207